dos/requirements.txt
scapy
python-socketio[client]
eventlet
psutil
sklearn
pandas
